pyproject.toml
[tool.poetry]
name = "warehouse_allocation"
version = "1.0.0"
description = "Genetic algorithm solver for warehouse allocation problem"
authors = ["Andrés Sandoval <andres.sndbrca@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
structlog = "20.2.0"
python-dotenv = "0.15.0"
pydantic = "1.7.3"
ortools = "9.0.9048"
numpy = "1.21.5"
pymoo = "0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"
docformatter = "^1.4"
pylint = "^2.6.0"
flake8 = "3.8.4"
mypy = "^0.800"
isort = "^5.7.0"
bandit = "^1.7.0"
black = "20.8b1"
jupyter = "^1.0.0"
matplotlib = "^3.3.4"
sphinx-rtd-theme = "^0.5.2"
sphinx-pydantic = "^0.1.1"
sphinxcontrib-bibtex = "^2.3.0"
sphinx-autodoc-typehints = "^1.12.0"
seed-intersphinx-mapping = "^0.5.1"
coverage = "^5.5"
nbsphinx = "^0.8.8"

[tool.black]
line-length = 79

[build-system]
requires = [
    "setuptools>=42",
    "wheel",
]
build-backend = "setuptools.build_meta"
